Refactor: Migrate WiFi state observation from platform-specific expect/actual functions to a CompositionLocal provided by MainActivity.
