--- compose-ui/src/androidMain/kotlin/dev/dimension/flare/ui/component/platform/PlatformWifiState.android.kt
@@ -1,48 +1,0 @@
-package dev.dimension.flare.ui.component.platform
-
-import android.content.Context
-import android.net.ConnectivityManager
-import android.net.Network
-import android.net.NetworkCapabilities
-import android.net.NetworkRequest
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.State
-import androidx.compose.runtime.produceState
-import androidx.compose.ui.platform.LocalContext
-import kotlinx.coroutines.channels.awaitClose
-import kotlinx.coroutines.flow.callbackFlow
-
-@Composable
-internal actual fun rememberPlatformWifiState(): State<Boolean> = wifiState()
-
-@Composable
-private fun wifiState(): State<Boolean> {
-    val context = LocalContext.current
-    return produceState(initialValue = false) {
-        context.observeWifiStateAsFlow().collect { value = it }
-    }
-}
-
-private fun Context.observeWifiStateAsFlow() =
-    callbackFlow {
-        val connectivityManager =
-            getSystemService(Context.CONNECTIVITY_SERVICE) as ConnectivityManager
-        val networkRequest =
-            NetworkRequest
-                .Builder()
-                .addCapability(NetworkCapabilities.NET_CAPABILITY_INTERNET)
-                .build()
-        val networkCallback =
-            object : ConnectivityManager.NetworkCallback() {
-                override fun onCapabilitiesChanged(
-                    network: Network,
-                    networkCapabilities: NetworkCapabilities,
-                ) {
-                    trySend(networkCapabilities.hasTransport(NetworkCapabilities.TRANSPORT_WIFI))
-                }
-            }
-        connectivityManager.registerNetworkCallback(networkRequest, networkCallback)
-        awaitClose {
-            connectivityManager.unregisterNetworkCallback(networkCallback)
-        }
-    }
