--- compose-ui/src/iosMain/kotlin/dev/dimension/flare/ui/component/platform/PlatformWifiState.ios.kt
@@ -1,7 +1,0 @@
-package dev.dimension.flare.ui.component.platform
-
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.State
-
-@Composable
-internal actual fun rememberPlatformWifiState(): State<Boolean> = androidx.compose.runtime.mutableStateOf(true)
