--- compose-ui/src/jvmMain/kotlin/dev/dimension/flare/ui/component/platform/PlatformWifiState.jvm.kt
@@ -1,10 +1,0 @@
-package dev.dimension.flare.ui.component.platform
-
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.State
-
-@Composable
-internal actual fun rememberPlatformWifiState(): State<Boolean> {
-    // TODO: Implement for JVM
-    return androidx.compose.runtime.remember { androidx.compose.runtime.mutableStateOf(true) }
-}
